#118 [REFACTOR] /src/api에 index.js 추가
